Renaming and refactoring
