+++ packages/core/src/data/ome_zarr/image_loader.ts
@@ -18,7 +18,7 @@ import { Image as OmeZarrImage } from "./0.5/image";
 
 import { Readable } from "@zarrita/storage";
 import { ZarrArrayParams } from "../zarr/open";
-import { getChunk } from "./zarr_worker_pool";
+import { getChunk } from "./worker_pool";
 
 // Implements the interface required for getting array chunks in zarrita:
 // https://github.com/manzt/zarrita.js/blob/c15c1a14e42a83516972368ac962ebdf56a6dcdb/packages/indexing/src/types.ts#L52
